pyproject.toml
[project]
name = "gameagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.2",
    "litellm>=1.74.7",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
]
